# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A production-ready Job Portal backend with role-based access control.",
        "title": "Job Portal API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@jobportal.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/jobs": {
            "get": {
                "security": [{ "BearerAuth": [] }],
                "description": "Get paginated list of jobs with optional location and company filters",
                "produces": ["application/json"],
                "tags": ["Jobs"],
                "summary": "List all jobs",
                "parameters": [
                    { "type": "integer", "default": 1, "description": "Page number", "name": "page", "in": "query" },
                    { "type": "integer", "default": 10, "description": "Items per page", "name": "page_size", "in": "query" },
                    { "type": "string", "description": "Filter by location", "name": "location", "in": "query" },
                    { "type": "string", "description": "Filter by company", "name": "company", "in": "query" }
                ],
                "responses": {
                    "200": { "description": "OK", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "401": { "description": "Unauthorized", "schema": { "$ref": "#/definitions/APIResponse" } }
                }
            },
            "post": {
                "security": [{ "BearerAuth": [] }],
                "description": "Recruiter creates a new job posting",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["Jobs"],
                "summary": "Create a job",
                "parameters": [
                    {
                        "description": "Job payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": { "$ref": "#/definitions/CreateJobInput" }
                    }
                ],
                "responses": {
                    "201": { "description": "Created", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "400": { "description": "Bad Request", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "401": { "description": "Unauthorized", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "403": { "description": "Forbidden", "schema": { "$ref": "#/definitions/APIResponse" } }
                }
            }
        },
        "/jobs/{id}": {
            "get": {
                "security": [{ "BearerAuth": [] }],
                "description": "Retrieve a single job by its UUID",
                "produces": ["application/json"],
                "tags": ["Jobs"],
                "summary": "Get a job by ID",
                "parameters": [
                    { "type": "string", "description": "Job UUID", "name": "id", "in": "path", "required": true }
                ],
                "responses": {
                    "200": { "description": "OK", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "400": { "description": "Bad Request", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "404": { "description": "Not Found", "schema": { "$ref": "#/definitions/APIResponse" } }
                }
            },
            "put": {
                "security": [{ "BearerAuth": [] }],
                "description": "Recruiter updates their own job posting",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["Jobs"],
                "summary": "Update a job",
                "parameters": [
                    { "type": "string", "description": "Job UUID", "name": "id", "in": "path", "required": true },
                    {
                        "description": "Update payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": { "$ref": "#/definitions/UpdateJobInput" }
                    }
                ],
                "responses": {
                    "200": { "description": "OK", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "400": { "description": "Bad Request", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "403": { "description": "Forbidden", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "404": { "description": "Not Found", "schema": { "$ref": "#/definitions/APIResponse" } }
                }
            },
            "delete": {
                "security": [{ "BearerAuth": [] }],
                "description": "Recruiter deletes their own job (Admin can delete any)",
                "produces": ["application/json"],
                "tags": ["Jobs"],
                "summary": "Delete a job",
                "parameters": [
                    { "type": "string", "description": "Job UUID", "name": "id", "in": "path", "required": true }
                ],
                "responses": {
                    "200": { "description": "OK", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "400": { "description": "Bad Request", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "403": { "description": "Forbidden", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "404": { "description": "Not Found", "schema": { "$ref": "#/definitions/APIResponse" } }
                }
            }
        },
...
        "/admin/users": {
            "get": {
                "security": [{ "BearerAuth": [] }],
                "description": "Admin retrieves all registered users with pagination",
                "produces": ["application/json"],
                "tags": ["Admin"],
                "summary": "[Admin] List all users",
                "parameters": [
                    { "type": "integer", "default": 1, "description": "Page number", "name": "page", "in": "query" },
                    { "type": "integer", "default": 10, "description": "Items per page", "name": "page_size", "in": "query" }
                ],
                "responses": {
                    "200": { "description": "OK", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "401": { "description": "Unauthorized", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "403": { "description": "Forbidden", "schema": { "$ref": "#/definitions/APIResponse" } }
                }
            }
        },
        "/admin/users/{id}": {
            "delete": {
                "security": [{ "BearerAuth": [] }],
                "description": "Admin soft-deletes a user by UUID",
                "produces": ["application/json"],
                "tags": ["Admin"],
                "summary": "[Admin] Delete a user",
                "parameters": [
                    { "type": "string", "description": "User UUID", "name": "id", "in": "path", "required": true }
                ],
                "responses": {
                    "200": { "description": "OK", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "400": { "description": "Bad Request", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "404": { "description": "Not Found", "schema": { "$ref": "#/definitions/APIResponse" } }
                }
            }
        },
        "/admin/jobs": {
            "get": {
                "security": [{ "BearerAuth": [] }],
                "description": "Admin retrieves all job postings with pagination and filters",
                "produces": ["application/json"],
                "tags": ["Admin"],
                "summary": "[Admin] List all jobs",
                "parameters": [
                    { "type": "integer", "default": 1, "description": "Page number", "name": "page", "in": "query" },
                    { "type": "integer", "default": 10, "description": "Items per page", "name": "page_size", "in": "query" },
                    { "type": "string", "description": "Filter by location", "name": "location", "in": "query" },
                    { "type": "string", "description": "Filter by company", "name": "company", "in": "query" }
                ],
                "responses": {
                    "200": { "description": "OK", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "401": { "description": "Unauthorized", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "403": { "description": "Forbidden", "schema": { "$ref": "#/definitions/APIResponse" } }
                }
            }
        },
...
    },
    "definitions": {
        "APIResponse": {
            "type": "object",
            "properties": {
                "success": { "type": "boolean" },
                "message": { "type": "string" },
                "data": {},
                "error": {}
            }
        },
...
        "CreateJobInput": {
            "type": "object",
            "required": ["title", "description", "company", "location"],
            "properties": {
                "title":       { "type": "string", "example": "Senior Go Engineer" },
                "description": { "type": "string", "example": "We are looking for an experienced Go developer." },
                "company":     { "type": "string", "example": "Acme Corp" },
                "location":    { "type": "string", "example": "New York, NY" }
            }
        },
        "UpdateJobInput": {
            "type": "object",
            "properties": {
                "title":       { "type": "string", "example": "Lead Go Engineer" },
                "description": { "type": "string", "example": "Updated description." },
                "company":     { "type": "string", "example": "Acme Corp" },
                "location":    { "type": "string", "example": "Remote" }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "description": "Enter your JWT token as: Bearer \u003ctoken\u003e"
        }
    }
}